# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_interfaces_http_handlers.CreateReminderRequest:
    properties:
      description:
        type: string
      recurrence:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_reminders.RecurrenceType'
        enum:
        - none
        - daily
        - weekly
        - monthly
        - custom
      recurrence_value:
        type: integer
      reminder_time:
        type: string
      title:
        type: string
    required:
    - recurrence
    - reminder_time
    - title
    type: object
...
  internal_interfaces_http_handlers.ErrorResponse:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: validation failed
        type: string
    type: object
...
  internal_interfaces_http_handlers.PresignRequest:
    properties:
      content_type:
        type: string
      file_name:
        type: string
      file_size:
        minimum: 1
        type: integer
    required:
    - content_type
    - file_name
    - file_size
    type: object
  internal_interfaces_http_handlers.PresignResponse:
    properties:
      expires_at:
        type: string
      file_url:
        type: string
      signed_id:
        type: string
      upload_url:
        type: string
    type: object
...
  internal_interfaces_http_handlers.ReminderListResponse:
    properties:
      count:
        example: 3
        type: integer
      reminders:
        items:
          $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.ReminderResponse'
        type: array
    type: object
...
  internal_interfaces_http_handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_interfaces_http_handlers.UpdateReminderRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      recurrence:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_reminders.RecurrenceType'
        enum:
        - none
        - daily
        - weekly
        - monthly
        - custom
      recurrence_value:
        type: integer
      reminder_time:
        type: string
      title:
        type: string
    type: object
...
  loviary_app_backend_internal_domain_reminders.RecurrenceType:
    enum:
    - none
    - daily
    - weekly
    - monthly
    - custom
    type: string
    x-enum-varnames:
    - RecurrenceNone
    - RecurrenceDaily
    - RecurrenceWeekly
    - RecurrenceMonthly
    - RecurrenceCustom
...
  loviary_app_backend_internal_interfaces_http_dto.ReminderResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_sent_at:
        type: string
      recurrence:
        $ref: '#/definitions/loviary_app_backend_internal_domain_reminders.RecurrenceType'
      recurrence_value:
        type: integer
      reminder_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@loviary.app
    name: API Support
  description: Loviary - Couple Relationship Tracking API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Loviary API
  version: "1.0"
paths:
...
  /reminders:
    get:
      consumes:
      - application/json
      description: Get all reminders for the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of reminders with count
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ReminderListResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all reminders
      tags:
      - reminders
    post:
      consumes:
      - application/json
      description: Create a new reminder with title, time and recurrence
      parameters:
      - description: Reminder data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Reminder created
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.ReminderResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create reminder
      tags:
      - reminders
  /reminders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a reminder by ID
      parameters:
      - description: Reminder ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reminder deleted
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.SuccessResponse'
        "400":
          description: Invalid reminder ID
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Reminder not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete reminder
      tags:
      - reminders
    get:
      consumes:
      - application/json
      description: Get reminder by ID
      parameters:
      - description: Reminder ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reminder data
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.ReminderResponse'
        "400":
          description: Invalid reminder ID
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Reminder not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get reminder
      tags:
      - reminders
    patch:
      consumes:
      - application/json
      description: Update an existing reminder
      parameters:
      - description: Reminder ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated reminder data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.UpdateReminderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated reminder
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.ReminderResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Reminder not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update reminder
      tags:
      - reminders
  /storage/presign:
    post:
      consumes:
      - application/json
      description: Get a presigned URL for uploading files to storage
      parameters:
      - description: Presign request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.PresignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Presigned URL and file info
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.PresignResponse'
        "400":
          description: Invalid input or file too large
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate presigned upload URL
      tags:
      - storage
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"